Refactor: extracted navigation into an extension class
